Added Resource Display to Main Screen.
After debugging ResourceDisplay.cs for a while I rewatched the tutorial (again) and saw that the time display function was in the game controller so I moved all of the resource code there and was able to get it to display on the main screen.
Default is to have only Resources display before the game starts and will display Money, Parts, and Morale after the game starts.
Since resources aren't consumed on every tic I currently have them updating every 10 tics to minimize overhead on updating the display.
