--- Assets/Scripts/ResourceDisplay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b72aa7e9a8eb2f45ab11219a29b41e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
